+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+ * `LedgerId`
+ * `Client.[set|get]LedgerId()`
+ * `TransferTransaction.addTokenTransferWithDecimals()`, `TransferTransaction.getDecimals()`.
+ * `ledgerId` fields in `AccountInfo`, `ContractInfo`, `FileInfo`, `ScheduleInfo`, `TokenInfo`, `TokenNftInfo`, and `TopicInfo`
+ * `UNEXPECTED_TOKEN_DECIMALS` response code.
+
+### Deprecated
+
+ * `NetworkName`, `Client.[set|get]NetworkName()`, user `LedgerId` and `Client.[set|get]LedgerId()` instead.
+
 ## v2.6.0-beta.2
 
 ### Fixed